Massive changes, Totally Function GUI
The GUI side of the calculator is working, using the Calc class for the calculations. The Run class was updated to only start up the calculator GUI. Most of the logic for the calculator to work according to certain conditions was implemented in the GUI class, where previously the command line calculator logic was implemented in the Run class.